Make caret transparent to mouse picking (issue 124) (#1032)
If you install a mouse handler on a TextExt area using setOnMouseClicked(), there's a small chance that your mouse click will be ignored (this happens approximately one time in ten in my experience).  The problem occurs when you click at the exact same position as the flashing caret while the caret is visible; this confuses the java FX mouse-picking algorithm, which causes the MouseEvent to be sent to the ParagraphText node rather than the TextExt. The fix is simply to make the caret transparent to mouse picking.

See this link for more evidence + screenshots : https://github.com/FXMisc/RichTextFX/issues/124